Test fixes and refactor
