+++ steps/accredited-programmes/login.ts
@@ -8,3 +8,12 @@ export const login = async (page: Page) => {
     await page.click('#submit')
     await expect(page).toHaveTitle(/Accredited Programmes/)
 }
+
+export const manageAndDeliverLogin = async (page: Page) => {
+    await page.goto(process.env.ACCREDITED_PROGRAMMES_MANAGE_AND_DELIVER_URL)
+    await expect(page).toHaveTitle(/HMPPS Digital Services - Sign in/)
+    await page.fill('#username', process.env.DELIUS_USERNAME!)
+    await page.fill('#password', process.env.DELIUS_PASSWORD!)
+    await page.click('#submit')
+    await expect(page).toHaveTitle(/Accredited Programmes/)
+}
